Specify gem `jar-dependencies` version
as this gem default version in jRuby is 0.4.1,
but the latest version is 0.5.1

Without a specific version, the following error will be raised on the CI servers:
Gem::LoadError: You have already activated jar-dependencies 0.4.1, but your Gemfile requires jar-dependencies 0.5.1. Since jar-dependencies is a default gem, you can either remove your dependency on it or try updating to a newer version of bundler that supports jar-dependencies as a default gem.